requirements.txt
absl-py==1.2.0
attrs==21.4.0
gym==0.21.0
idna==2.7
imageio==2.20.0
importlib-metadata==4.12.0
lbforaging==1.1.1
matplotlib==3.5.2
mpyq==0.2.5
numpy==1.21.6
pbr==4.3.0
Pillow==9.2.0
pluggy==0.7.1
portpicker==1.2.0
probscale==0.2.5
protobuf==3.19.4
py==1.6.0
pygame==2.1.2
pyparsing==2.2.2
pysc2==3.0.0
pytest==4.3.1
python-dateutil==2.7.3
PyYAML==5.3.1
requests==2.28.1
rware==1.0.3
s2clientprotocol==4.10.1.75800.0
sacred==0.8.2
scipy==1.7.3
six==1.11.0
sk-video==1.1.10
snakeviz==2.1.1
tensorboard-logger==0.1.0
torch==1.12.0
torchvision==0.13.0
tornado==5.1.1
urllib3==1.23
websocket-client==0.53.0
whichcraft==0.5.2
wrapt==1.10.11
tianshou==0.4.3
pymoo==0.5.0
geatpy==2.6.0
ray==1.9.0
platypus-opt==1.0.4